+++ package.json
@@ -29,7 +29,7 @@
     "e2e-ci": "cross-env BABEL_ENV=test CYPRESS_USE_FIXTURES=true cypress run --record --parallel",
     "vercel-build": "npm run build",
     "vercel-start": "cross-env NODE_ENV=production NOW=true node build/server | bunyan",
-    "get-gql-schema-local": "rm src/schema.graphql && apollo client:download-schema --endpoint=http://localhost:4000/graphql-api/graphql src/schema.graphql && prettier --write ./src/schema.graphql",
+    "get-gql-schema-local": "rm src/schema.graphql -f && apollo client:download-schema --endpoint=http://localhost:4000/graphql-api/graphql src/schema.graphql && prettier --write ./src/schema.graphql",
     "generate-gql-types": "graphql-codegen --config codegen.yml && yarn format",
     "check-translations": "cross-env BABEL_ENV=test NODE_ENV=unittest jest src/messages/__tests__/"
   },
@@ -142,7 +142,6 @@
     "react-dom": "^17.0.2",
     "react-helmet-async": "^1.3.0",
     "react-i18next": "11.11.4",
-    "react-query": "^3.39.1",
     "react-router-dom": "^6.3.0",
     "remarkable": "^2.0.0",
     "serialize-javascript": "^3.1.0",
